requirements.txt
streamlit
openai-whisper
yt-dlp
pytube

